+++ src/themes/raw/gsi/parse-players.js
@@ -56,17 +56,16 @@ const parsePlayerWeapons = (player) => Object.values(player.weapons).flatMap((we
 export const parsePlayers = () => {
 	const playerNameOverrides = getPlayerNameOverrides()
 	const { hiddenPlayerNames, hiddenPlayerSteam64Ids } = getHiddenPlayers()
+	const hideCoachesByName = options['teams.hideCoachesByName']
 
 	const players = []
-	const autoHideCoaches = options['teams.autoHideCoaches']
 
 	for (const [steam64Id, player] of Object.entries(gsiState.allplayers)) {
 		if (hiddenPlayerSteam64Ids.has(steam64Id)) continue
 
 		const name = playerNameOverrides.get(steam64Id) || player.name
 		if (hiddenPlayerNames.has(name)) continue
-
-		if (autoHideCoaches && /coach/i.test(name)) continue
+		if (hideCoachesByName && /\bcoach\b/i.test(name)) continue
 
 		const observerSlot = getObserverSlot(player, steam64Id)
 		if (observerSlot === undefined) continue